[CHORE] #53 : signin, singup -> login, register 수정, 기존 클래스 네임에서 sign 삭제
