+++ client/app/src/main/java/com/example/healthc/presentation/auth/register/UserEmailFragment.kt
@@ -1,4 +1,4 @@
-package com.example.healthc.presentation.auth.sign_up
+package com.example.healthc.presentation.auth.register
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -11,17 +11,16 @@ import androidx.lifecycle.flowWithLifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
 import com.example.healthc.R
-import com.example.healthc.databinding.FragmentSignUpPasswordBinding
+import com.example.healthc.databinding.FragmentUserEmailBinding
 import com.example.healthc.presentation.auth.AuthViewModel
-import com.example.healthc.presentation.auth.AuthViewModel.SignUpUiEvent
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 
 @AndroidEntryPoint
-class SignUpPasswordFragment : Fragment() {
+class UserEmailFragment : Fragment() {
 
-    private var _binding: FragmentSignUpPasswordBinding? = null
+    private var _binding: FragmentUserEmailBinding? = null
     private val binding get() = checkNotNull(_binding)
 
     private val viewModel by activityViewModels<AuthViewModel>()
@@ -31,7 +30,7 @@ class SignUpPasswordFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = DataBindingUtil.inflate(inflater, R.layout.fragment_sign_up_password, container, false)
+        _binding = DataBindingUtil.inflate(inflater, R.layout.fragment_user_email, container, false)
         binding.viewModel = viewModel
         binding.lifecycleOwner = viewLifecycleOwner
         return binding.root
@@ -46,21 +45,21 @@ class SignUpPasswordFragment : Fragment() {
         viewModel.signUpUiEvent.flowWithLifecycle(viewLifecycleOwner.lifecycle)
             .onEach {
                 when (it) {
-                    is SignUpUiEvent.Success -> {
-                        navigateToInfo()
+                    is AuthViewModel.SignUpUiEvent.Success -> {
+                        navigateToPassword()
                         viewModel.initState()
                     }
-                    is SignUpUiEvent.Failure -> {
-                        binding.signUpPasswordEditView.error = it.message
+                    is AuthViewModel.SignUpUiEvent.Failure -> {
+                        binding.signUpEmailLayout.error = it.message
                     }
-                    is SignUpUiEvent.Unit -> {}
+                    is AuthViewModel.SignUpUiEvent.Unit -> {}
                 }
             }.launchIn(viewLifecycleOwner.lifecycleScope)
     }
 
-    private fun navigateToInfo() {
+    private fun navigateToPassword() {
         lifecycleScope.launchWhenStarted {
-            val direction = SignUpPasswordFragmentDirections.actionSignUpPasswordFragmentToSignUpInfoFragment()
+            val direction = UserEmailFragmentDirections.actionUserEmailFragmentToUserPasswordFragment()
             findNavController().navigate(direction)
         }
     }
