+++ client/app/src/main/java/com/example/healthc/presentation/auth/register/UserPasswordFragment.kt
@@ -1,11 +1,9 @@
-package com.example.healthc.presentation.auth.sign_in
+package com.example.healthc.presentation.auth.register
 
-import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Toast
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
@@ -13,18 +11,17 @@ import androidx.lifecycle.flowWithLifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
 import com.example.healthc.R
-import com.example.healthc.databinding.FragmentSignInBinding
-import com.example.healthc.domain.utils.Resource
+import com.example.healthc.databinding.FragmentUserPasswordBinding
 import com.example.healthc.presentation.auth.AuthViewModel
-import com.example.healthc.presentation.main.MainActivity
+import com.example.healthc.presentation.auth.AuthViewModel.SignUpUiEvent
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 
 @AndroidEntryPoint
-class SignInFragment : Fragment() {
+class UserPasswordFragment : Fragment() {
 
-    private var _binding: FragmentSignInBinding? = null
+    private var _binding: FragmentUserPasswordBinding? = null
     private val binding get() = checkNotNull(_binding)
 
     private val viewModel by activityViewModels<AuthViewModel>()
@@ -34,7 +31,7 @@ class SignInFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = DataBindingUtil.inflate(inflater, R.layout.fragment_sign_in, container, false)
+        _binding = DataBindingUtil.inflate(inflater, R.layout.fragment_user_password, container, false)
         binding.viewModel = viewModel
         binding.lifecycleOwner = viewLifecycleOwner
         return binding.root
@@ -43,45 +40,29 @@ class SignInFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         observeData()
-        initView()
     }
 
     private fun observeData(){
-        viewModel.signInEvent.flowWithLifecycle(viewLifecycleOwner.lifecycle)
+        viewModel.signUpUiEvent.flowWithLifecycle(viewLifecycleOwner.lifecycle)
             .onEach {
-                when(it){
-                    is Resource.Loading ->{
-                        // TODO loading screen
+                when (it) {
+                    is SignUpUiEvent.Success -> {
+                        navigateToInfo()
+                        viewModel.initState()
                     }
-                    is Resource.Success -> {
-                        startMainActivity()
+                    is SignUpUiEvent.Failure -> {
+                        binding.signUpPasswordEditView.error = it.message
                     }
-                    is Resource.Failure -> {
-                        Toast.makeText(requireContext(), "로그인에 실패하였습니다.", Toast.LENGTH_SHORT).show()
-                    }
-                    else -> {}
+                    is SignUpUiEvent.Unit -> {}
                 }
             }.launchIn(viewLifecycleOwner.lifecycleScope)
     }
 
-    private fun initView(){
-        binding.goToSignUpButton.setOnClickListener {
-            viewModel.initInput()
-            navigateSignUp()
-        }
-    }
-
-    private fun navigateSignUp(){
+    private fun navigateToInfo() {
         lifecycleScope.launchWhenStarted {
-            val direction = SignInFragmentDirections.actionSignInFragmentToSignUpNameFragment()
+            val direction = UserPasswordFragmentDirections.actionUserPasswordFragmentToUserInfoFragment()
             findNavController().navigate(direction)
         }
-    }
-
-    private fun startMainActivity(){
-        val intent = Intent(requireContext(), MainActivity::class.java)
-        intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
-        startActivity(intent)
     }
 
     override fun onDestroyView() {
