+++ client/app/src/main/res/layout/fragment_user_password.xml
